Loop through all songs and check with Shazam
We now look up each song in the playlist on Shazam.
There are two modes: Quick (use the file on disk) and if that doesn't match, we go to Deep
Deep is where we stream the song then check.  For some reason the Shazam API behaves better with a stream than with the file.

Resolves #8
Resolves #9
Resolves #10